playgroundCelery/requirements.txt
amqp==2.5.0
attrs==19.1.0
backcall==0.1.0
billiard==3.6.0.0
bleach==3.1.0
celery==4.3.0
decorator==4.4.0
defusedxml==0.6.0
Django==2.2
django-celery-results==1.1.2
djangorestframework==3.9.4
entrypoints==0.3
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kombu==4.6.1
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
pandocfilters==1.4.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.4.2
pyrsistent==0.15.2
python-dateutil==2.8.0
pytz==2019.1
pyzmq==18.0.1
qtconsole==4.5.1
redis==3.2.1
Send2Trash==1.5.0
six==1.12.0
sqlparse==0.3.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
traitlets==4.3.2
vine==1.3.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
